[IMP] runbot: add a branch name_search that reflects the name_get
When creating a cronbuild, the name that is displayed also contains the
repository name. That way, it's easier to select the right branch.
But when the users starts typing, the search filters only on the branch
name.

With this commit, the name_search reflects the name_get. That way, when
the user starts typing, it filters on the repo name or the branch name.
After a ':' is typed, it filters the first part on the repository name and
the second part on the branch name.